Revert "fix some issues where multiplication result is a larger type than the factors"
This reverts commit 923888e45074536bfd426f56aa997d418bf98d6c.

+++ Src/JxrDecode/Jxr/JXRGlue.c
@@ -664,7 +664,7 @@ ERR PKImageEncode_WriteSource(
     cbStride = MAX(cbStrideFrom, cbStrideTo);
 
     // actual dec/enc with local buffer
-    Call(PKAllocAligned((void **) &pb, (size_t)cbStride * pRect->Height, 128));
+    Call(PKAllocAligned((void **) &pb, cbStride * pRect->Height, 128));
 
     Call(pFC->Copy(pFC, pRect, pb, cbStride));
 
@@ -765,7 +765,7 @@ ERR PKImageEncode_Transcode(
 	else 
 	{
 		// actual dec/enc with local buffer
-	    Call(PKAllocAligned((void **) &pb, (size_t)cbStride * pRect->Height, 128));
+	    Call(PKAllocAligned((void **) &pb, cbStride * pRect->Height, 128));
 		Call(pFC->Copy(pFC, pRect, pb, cbStride));
 		Call(pIE->WritePixels(pIE, pRect->Height, pb, cbStride));
 	}
